Move delete into application service
